Add freezeWhen to the email field
