+++ components/brave_account/resources/brave_account_password_input.ts
@@ -3,12 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import {
-  AsyncDirective,
-  CrLitElement,
-  directive,
-  nothing,
-} from '//resources/lit/v3_0/lit.rollup.js'
+import { CrLitElement } from '//resources/lit/v3_0/lit.rollup.js'
 
 import { getCss } from './brave_account_common.css.js'
 import { getHtml } from './brave_account_password_input.html.js'
@@ -20,20 +15,6 @@ import type { ToggleVisibilityEventDetail } from './brave_account_password_icons
 // up to 32kB with characters outside the BMP, e.g. emojis). This limits
 // length in code units, not bytes.
 export const MAX_PASSWORD_LENGTH = 8192
-
-// Custom directive that freezes the previously rendered value when `freeze`
-// is true. Similar to Lit's `noChange` (not exported by Chromium's Lit wrapper
-// from //third_party/lit/v3_0/lit.ts), but instead of preventing the update,
-// it reuses the last rendered value.
-class FreezeWhenDirective extends AsyncDirective {
-  private previousValue: unknown = nothing
-
-  render(freeze: boolean, value: unknown): unknown {
-    return freeze ? this.previousValue : (this.previousValue = value)
-  }
-}
-
-export const freezeWhen = directive(FreezeWhenDirective)
 
 export type PasswordInputEventDetail = { password: string; isValid: boolean }
